First attempt at finding all photos not in a shared album.
This gets all shared albums, gets all photos in each of those shared albums, gets all photos, and then uses that to find the photos not in shared albums. It then attempts to use the Photos Library API to create an album with those photos, but that doesn't work because the photos weren't create by this "app" (https://stackoverflow.com/questions/56374316/google-photo-returns-error-400-request-contains-an-invalid-media-item-id-inv).

This also moves the secrets to an env file, gitignores it, and adds a script to run the server with it.